Example outbreak data for Hepatatis E virus or Acute Jaundice Syndrome (AJS)

Data from a real outbreak of Hepatitis E virus (HEV) infection which occurred in the Chadian town of Am Timan between October 2016 and April 2017. The Chadian Ministry of Health (MoH) has approved the use of this data for training purposes. Please note, that some data has been adapted in order to best achieve training objectives. Also, the GPS coordinates included in the dataset do not correspond to real cases identified during this outbreak. They have been generated exclusively for training purposes.

data(AJS_linelist_internal)

Format

An object of class "cross"; see read.cross.

Source

Article

References

None (PLoS)

Examples

data(AJS_linelist_internal)
head(AJS_linelist_internal)
#>   Number HEVID Sex Age Agegroup  dateentry epiquestdate team  chw block   hh
#> 1      1  1001   F  35        3 2016-09-01         <NA> <NA> <NA>  <NA> <NA>
#> 2      2  1002   F  21        3 2016-09-06         <NA> <NA> <NA>  <NA> <NA>
#> 3      3  1003   F  18        3 2016-09-09         <NA> <NA> <NA>  <NA> <NA>
#> 4      4  1004   F  25        3 2016-09-12   2016-09-12 <NA> <NA>  <NA> <NA>
#> 5      5  1005   F  20        3 2016-09-19   2016-09-27 <NA> <NA>  <NA> <NA>
#> 6      6  1006   M   8        2 2016-09-19         <NA> <NA> <NA>  <NA> <NA>
#>   referredby   quartier interviewee hhpers hhmale hhfemale hhu5 hpreferral
#> 1       <NA>  Rassalfil        <NA>   <NA>   <NA>     <NA> <NA>       <NA>
#> 2       <NA>  Anfandock        <NA>   <NA>   <NA>     <NA> <NA>       <NA>
#> 3       <NA>  Al-Hougna        <NA>   <NA>   <NA>     <NA> <NA>       <NA>
#> 4       <NA>   Dinguess      Parent   <NA>   <NA>     <NA> <NA>       <NA>
#> 5       <NA> Taradona 2     Patient      3      2        1    1       <NA>
#> 6       <NA>   Albayadi        <NA>   <NA>   <NA>     <NA> <NA>       <NA>
#>   hprefpos patientpresent ptjaundice ptvomit mtmental ptpregnant ptu1 referring
#> 1     <NA>           <NA>       <NA>    <NA>     <NA>       <NA> <NA>      <NA>
#> 2     <NA>           <NA>       <NA>    <NA>     <NA>       <NA> <NA>      <NA>
#> 3     <NA>           <NA>       <NA>    <NA>     <NA>       <NA> <NA>      <NA>
#> 4     <NA>           <NA>       <NA>    <NA>     <NA>       <NA> <NA>      <NA>
#> 5     <NA>           <NA>       <NA>    <NA>     <NA>       <NA> <NA>      <NA>
#> 6     <NA>           <NA>       <NA>    <NA>     <NA>       <NA> <NA>      <NA>
#>   otherajs otherajsnum borehole comtap river citywater otherwater otherwatersp
#> 1     <NA>        <NA>     <NA>   <NA>  <NA>      <NA>       <NA>         <NA>
#> 2     <NA>        <NA>     <NA>   <NA>  <NA>      <NA>       <NA>         <NA>
#> 3     <NA>        <NA>     <NA>   <NA>  <NA>      <NA>       <NA>         <NA>
#> 4     <NA>        <NA>     <NA>    Oui  <NA>      <NA>       <NA>         <NA>
#> 5      Non        <NA>      Oui   <NA>  <NA>      <NA>       <NA>         <NA>
#> 6     <NA>        <NA>     <NA>   <NA>  <NA>      <NA>       <NA>         <NA>
#>   watersourcefreq ceramicjar jerrycan oildrum reserother reserothersp resertype
#> 1            <NA>       <NA>     <NA>    <NA>       <NA>         <NA>      <NA>
#> 2            <NA>       <NA>     <NA>    <NA>       <NA>         <NA>      <NA>
#> 3            <NA>       <NA>     <NA>    <NA>       <NA>         <NA>      <NA>
#> 4            <NA>        Oui      Oui    <NA>       <NA>         <NA>      <NA>
#> 5            <NA>        Oui     <NA>    <NA>       <NA>         <NA>      <NA>
#> 6            <NA>       <NA>     <NA>    <NA>       <NA>         <NA>      <NA>
#>   chlorination volumreser handreser hweating hwwater hwsoap treatwater
#> 1         <NA>       <NA>      <NA>     <NA>    <NA>   <NA>       <NA>
#> 2         <NA>       <NA>      <NA>     <NA>    <NA>   <NA>       <NA>
#> 3         <NA>       <NA>      <NA>     <NA>    <NA>   <NA>       <NA>
#> 4         <NA>       <NA>      <NA>     <NA>    <NA>   <NA>        Non
#> 5         <NA>       <NA>      <NA>     <NA>    <NA>   <NA>        Non
#> 6         <NA>       <NA>      <NA>     <NA>    <NA>   <NA>       <NA>
#>   tptreatwater tptreatwaterother                       nottreatwhy defecriver
#> 1         <NA>              <NA>                              <NA>       <NA>
#> 2         <NA>              <NA>                              <NA>       <NA>
#> 3         <NA>              <NA>                              <NA>       <NA>
#> 4         <NA>              <NA>                              <NA>       <NA>
#> 5         <NA>              <NA> L'EAU C'EST TRAITE DANS LE FORAGE       <NA>
#> 6         <NA>              <NA>                              <NA>       <NA>
#>   defecair defeclatrine numlatrine  dtmedical hpid outreachscreened
#> 1     <NA>         <NA>       <NA> 2016-09-01  633             <NA>
#> 2     <NA>         <NA>       <NA> 2016-09-06  643             <NA>
#> 3     <NA>         <NA>       <NA> 2016-09-09  649             <NA>
#> 4     <NA>          Oui       <NA> 2016-09-12  656                N
#> 5     <NA>          Oui       <NA> 2016-09-19   72                N
#> 6     <NA>         <NA>       <NA> 2016-09-19 1875             <NA>
#>   evaljaundice jaundiceid referredbychw jaundiceid2 medjaundice dtjaundice
#> 1         <NA>       <NA>          <NA>        <NA>        <NA>       <NA>
#> 2         <NA>       <NA>          <NA>        <NA>        <NA>       <NA>
#> 3         <NA>       <NA>          <NA>        <NA>        <NA>       <NA>
#> 4         <NA>       <NA>          <NA>        <NA>        <NA>       <NA>
#> 5         <NA>       <NA>          <NA>        <NA>        <NA> 2016-09-15
#> 6         <NA>       <NA>          <NA>        <NA>        <NA> 2016-09-14
#>   medfever mednausea medvomit medepigastric meditching medheadache
#> 1      Oui       Non      Non           Non        Non         Oui
#> 2      Oui       Oui      Oui           Oui        Non         Oui
#> 3      Non       Oui      Oui           Oui        Non         Oui
#> 4      Non       Oui      Non           Oui        Non         Oui
#> 5      Oui       Oui      Oui           Oui        Non         Oui
#> 6      Oui       Oui      Non           Oui        Non         Oui
#>   medarthralgia meddiar medbleeding medfev38        medmental medother
#> 1           Non     Oui         Non      Non           Normal     <NA>
#> 2           Oui     Oui         Non      Oui           Normal     <NA>
#> 3           Non     Non         Oui      Oui Confus/somnolent     <NA>
#> 4           Non     Non         Non      Non Confus/somnolent     <NA>
#> 5           Oui     Non         Non      Non           Normal     <NA>
#> 6           Oui     Oui         Non      Oui           Normal     <NA>
#>   medothhhajs medothhhajsnum othercaseid medpreg medpregtri medpp medppdays
#> 1         Non           <NA>        <NA>     Non       <NA>   Oui         5
#> 2     Unknown           <NA>        <NA>     Oui          1   Non      <NA>
#> 3         Non           <NA>        <NA>     Oui          3   Non      <NA>
#> 4         Non           <NA>        <NA>     Oui          2   Non      <NA>
#> 5         Non           <NA>        <NA>     Non       <NA>   Non      <NA>
#> 6         Oui           <NA>        <NA>     Non       <NA>   Non      <NA>
#>                        medppoutcome matdeathdt medppga medblood
#> 1 Naissance, enfant toujours vivant         NA    <NA>      Non
#> 2                  Sortie, enceinte         NA    <NA>      Non
#> 3 Naissance, enfant toujours vivant         NA    <NA>      Non
#> 4                Avortement/mort.né         NA      30      Oui
#> 5                              <NA>         NA       -      Oui
#> 6                              <NA>         NA    <NA>      Oui
#>         medblooddt medmalrdt medhepb medhepc medhevrdt medhevelisa hevrecent
#> 1               No       Neg    <NA>    <NA>      <NA>        <NA>      <NA>
#> 2               No       Pos    <NA>    <NA>      <NA>        <NA>      <NA>
#> 3 Yes but not sent       Neg     Neg     Neg       Neg        <NA>      <NA>
#> 4       2016-09-14       Neg     Neg     Neg       Pos   IgG+/IgM+         1
#> 5       2016-09-19       Neg     Neg     Neg       Neg   IgG-/IgM-         0
#> 6       2016-09-19       Pos     Neg     Neg       Pos   IgG+/IgM+         1
#>   hevpcr hevgenotest hevgenotype medhavelisa medhbvelisa medhcvelisa arbovpcr
#> 1   <NA>        <NA>        <NA>        <NA>        <NA>        <NA>     <NA>
#> 2   <NA>        <NA>        <NA>        <NA>        <NA>        <NA>     <NA>
#> 3   <NA>        <NA>        <NA>        <NA>        <NA>        <NA>     <NA>
#> 4    Pos        <NA>        <NA>         Neg         Neg         Neg     <NA>
#> 5    Neg        <NA>        <NA>         Neg         Neg         Neg      Neg
#> 6    Neg        <NA>        <NA>         Neg         Neg         Neg     <NA>
#>          yf      zika    dengue       wnv       rvf      chik hospitalised
#> 1      <NA>      <NA>      <NA>      <NA>      <NA>      <NA>          Oui
#> 2      <NA>      <NA>      <NA>      <NA>      <NA>      <NA>          Oui
#> 3      <NA>      <NA>      <NA>      <NA>      <NA>      <NA>          Oui
#> 4      <NA>      <NA>      <NA>      <NA>      <NA>      <NA>          Oui
#> 5 IgG-/IgM- IgG-/IgM- IgG-/IgM- IgG+/IgM- IgG-/IgM- IgG-/IgM-          Non
#> 6      <NA>      <NA>      <NA>      <NA>      <NA>      <NA>          Oui
#>   dthospitalisation      outcomehp    dtdeath epiwajs epiwhevelisa epiwoutreach
#> 1        2016-09-01 Déchargé/Guéri       <NA>    <NA>         <NA>      2016_35
#> 2        2016-09-06 Déchargé/Guéri       <NA>    <NA>         <NA>      2016_36
#> 3        2016-09-09         Décédé 2016-09-13    <NA>         <NA>      2016_36
#> 4        2016-09-12         Décédé 2016-09-14    <NA>         <NA>      2016_37
#> 5              <NA>           <NA>       <NA> 2016_37         <NA>      2016_38
#> 6        2016-09-19 Déchargé/Guéri       <NA> 2016_37      2016_37      2016_38
#>   epiwmed epiwepiquest  epiwhp epiwdeath  caseclass Latitude Longitude
#> 1 2016_35         <NA> 2016_35      <NA>    Suspect       NA        NA
#> 2 2016_36         <NA> 2016_36      <NA>    Suspect       NA        NA
#> 3 2016_36         <NA> 2016_36   2016_37 Not a case       NA        NA
#> 4 2016_37      2016_37 2016_37   2016_37  Confirmed 11.01991  20.28088
#> 5 2016_38      2016_39    <NA>      <NA> Not a case 11.05434  20.27419
#> 6 2016_38         <NA> 2016_38      <NA>  Confirmed       NA        NA