Display the available sitrep templates

Display the available sitrep templates

available_sitrep_templates(categorise = FALSE, ...)

Arguments

categorise
if TRUE, the results are split into a list of outbreak and survey categories. Defaults to FALSE.

...
options to pass on to dir

Value

a vector of available templates in the sitrep package

Examples

available_sitrep_templates(categorise = TRUE)
#> $outbreak
#> [1] "ajs_outbreak"        "cholera_outbreak"    "measles_outbreak"   
#> [4] "meningitis_outbreak"
#> 
#> $survey
#> [1] "mortality"   "nutrition"   "vaccination"
#> 
available_sitrep_templates(categorise = TRUE, full.names = TRUE)
#> $outbreak
#> [1] "/Users/runner/work/_temp/Library/sitrep/rmarkdown/templates/ajs_outbreak"       
#> [2] "/Users/runner/work/_temp/Library/sitrep/rmarkdown/templates/cholera_outbreak"   
#> [3] "/Users/runner/work/_temp/Library/sitrep/rmarkdown/templates/measles_outbreak"   
#> [4] "/Users/runner/work/_temp/Library/sitrep/rmarkdown/templates/meningitis_outbreak"
#> 
#> $survey
#> [1] "/Users/runner/work/_temp/Library/sitrep/rmarkdown/templates/mortality"  
#> [2] "/Users/runner/work/_temp/Library/sitrep/rmarkdown/templates/nutrition"  
#> [3] "/Users/runner/work/_temp/Library/sitrep/rmarkdown/templates/vaccination"
#>