Quartier shapefile for example outbreak Hepatatis E virus or Acute Jaundice Syndrome (AJS)

Shapefile for blocks (polygons) from a real outbreak of Hepatitis E virus (HEV) infection which occurred in the Chadian town of Am Timan between October 2016 and April 2017. The Chadian Ministry of Health (MoH) has approved the use of this data for training purposes. Please note, that some data has been adapted in order to best achieve training objectives. Also, the GPS coordinates included in the dataset do not correspond to real cases identified during this outbreak. They have been generated exclusively for training purposes.

data(block_shape)

Format

An object of class "cross"; see read.cross.

Source

Article

References

None (PLoS)

Examples

data(block_shape)
head(block_shape)
#> Simple feature collection with 6 features and 2 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 20.26417 ymin: 11.05669 xmax: 20.29393 ymax: 11.06318
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 3
#>   Name      Descriptio                                                  geometry
#>   <chr>     <chr>                                                  <POLYGON [°]>
#> 1 Block 369 NA         ((20.28043 11.05736, 20.27976 11.05845, 20.28049 11.0586…
#> 2 Block 617 NA         ((20.26849 11.05708, 20.26768 11.05806, 20.2677 11.05806…
#> 3 Block 631 NA         ((20.26512 11.05995, 20.26417 11.06117, 20.26468 11.0616…
#> 4 Block 646 NA         ((20.26692 11.06131, 20.26434 11.06259, 20.26462 11.0631…
#> 5 Block 411 NA         ((20.29184 11.05773, 20.29226 11.05781, 20.29235 11.0575…
#> 6 Block 412 NA         ((20.292 11.0583, 20.29349 11.05888, 20.29393 11.05818, …