Run the templates found in the sitrep package.

Run the templates found in the sitrep package.

check_sitrep_templates(
  templates = available_sitrep_templates(),
  path = tempdir(),
  quiet = FALSE,
  progress = FALSE,
  mustwork = FALSE,
  output_format = NULL,
  clean = TRUE
)

Arguments

templates
a vector of templates to create and render

path
a directory in which to store the templates

quiet
if FALSE (default), a message will be printed for every template.

progress
if TRUE, then the progress of the template is printed to screen (passed to the quiet argument of rmarkdown::render()). Defaults to FALSE

mustwork
if TRUE, then the templates must work for the function to succeed. Defaults to FALSE, which will simply print the errors.

output_format
a character defining the output formats to use for the template files. Defaults to the output_format defined in the templates (which is word_document), but can be modified to html_document for cross-platform cromulence checking.

clean
if TRUE (default), this will remove the previous output file before rendering.

Value

the path where the templates were built.

Examples

if (FALSE) {
check_sitrep_templates("mortality")
}