Fake spatial data as polygons This function returns a polygon which is split in to regions based on a supplied vector of names

Fake spatial data as polygons This function returns a polygon which is split in to regions based on a supplied vector of names

gen_polygon(regions)

Arguments

regions
A string of names for each region to label the polygon with

References

The coordinates used for the polygon are of Vienna, Austria. based off government data (see metadata)