MSF data dictionaries and dummy datasets

These function reads in MSF data dictionaries and produces randomised datasets based on values defined in the dictionaries. The randomised dataset produced should mimic an excel export from DHIS2.

msf_dict(disease, name = "MSF-outbreak-dict.xlsx", tibble = TRUE,
  compact = TRUE, long = TRUE)

msf_dict_rename_helper(disease, varnames = "data_element_shortname",
  copy_to_clipboard = TRUE)

gen_data(dictionary, varnames = "data_element_shortname",
  numcases = 300)

msf_dict_survey(disease, name = "MSF-survey-dict.xlsx", tibble = TRUE,
  compact = FALSE)

Arguments

disease
Specify which disease you would like to use. Currently supports "Cholera", "Measles" and "Meningitis".

name
the name of the dictionary stored in the package.

tibble
Return data dictionary as a tidyverse tibble (default is TRUE)

compact
if TRUE (default), then a nested data frame is returned where each row represents a single variable and a nested data frame column called "options", which can be expanded with tidyr::unnest(). This only works if long = TRUE.

long
If TRUE (default), the returned data dictionary is in long format with each option getting one row. If FALSE, then two data frames are returned, one with variables and the other with content options.

varnames
Specify name of column that contains varnames. Currently default set to "Item". (this can probably be deleted once dictionaries standardise) If dictionary is "Mortality", varnames needs to be "column_name"`.

copy_to_clipboard
if TRUE (default), the rename template will be copied to the user's clipboard with clipr::write_clip(). If FALSE, the rename template will be printed to the user's console.

dictionary
Specify which dictionary you would like to use. Currently supports "Cholera", "Measles", "Meningitis", "AJS" and "Mortality".

numcases
For fake data, specify the number of cases you want (default is 300

See also

Examples

if (require('dplyr') & require('linelist')) { withAutoprint({
# You will often want to use MSF dictionaries to translate codes to human-
# readable variables. Here, we generate a data set of 20 cases:
dat <- gen_data(dictionary = "Cholera", varnames = "data_element_shortname",
                numcases = 20)
print(dat)

# We want the expanded dictionary, so we will select `compact = FALSE`
dict <- msf_dict(disease = "Cholera", long = TRUE, compact = FALSE, tibble = TRUE)
print(dict)

# We can use linelist's clean_variable_spelling to translate the codes. First,
# we want to reorder the columns of the dictionary like so:
#
#  - 1st column: option codes
#  - 2nd column: translations
#  - 3rd column: data column name
#  - 4th column: order of options
#
# we also want to make sure to filter out any columns that are blank for
# the option codes, because this means that they don't have a fixed number of
# options
dict <- dict %>%
  select(option_code, option_name, data_element_shortname, option_order_in_set) %>%
  filter(!is.na(option_code))
print(dict)

# Now we can use linelist to filter the data:
dat_clean <- clean_variable_spelling(dat, dict)
print(dat_clean)
})}

#> Loading required package: linelist

#> > dat <- gen_data(dictionary = "Cholera", varnames = "data_element_shortname", 
#> +     numcases = 20)
#> > print(dat)
#> # A tibble: 20 x 45
#>    trimester exit_status treatment_facil… dehydration_lev… comments_on_lab…
#>    <fct>     <fct>       <lgl>            <fct>            <lgl>           
#>  1 NA        LA          NA               SO               NA              
#>  2 NA        DOA         NA               SE               NA              
#>  3 NA        DH          NA               SO               NA              
#>  4 NA        DD          NA               UN               NA              
#>  5 NA        DOA         NA               SE               NA              
#>  6 NA        AD          NA               NO               NA              
#>  7 NA        DOA         NA               NO               NA              
#>  8 NA        AD          NA               UN               NA              
#>  9 NA        AD          NA               SO               NA              
#> 10 NA        AD          NA               UN               NA              
#> 11 NA        TR          NA               SE               NA              
#> 12 NA        LA          NA               UN               NA              
#> 13 NA        LA          NA               SO               NA              
#> 14 NA        DD          NA               NO               NA              
#> 15 NA        AD          NA               UN               NA              
#> 16 NA        DH          NA               NO               NA              
#> 17 NA        TR          NA               NO               NA              
#> 18 NA        AD          NA               SO               NA              
#> 19 NA        DH          NA               SO               NA              
#> 20 NA        DD          NA               UN               NA              
#> # … with 40 more variables: fluids_treatment_plan <fct>, time_to_death <fct>,
#> #   malaria_rdt_at_admission <fct>, previously_vaccinated <fct>,
#> #   cholera_pcr_result <fct>, iv_fluids_received_litres <int>,
#> #   patient_origin_free_text <chr>, previous_vaccine_doses_received <fct>,
#> #   age_years <int>, arrival_date_in_area_if_3m <date>, age_days <lgl>,
#> #   msf_involvement <fct>, date_lab_sample_taken <date>, readmission <fct>,
#> #   age_months <int>, date_of_last_vaccination <date>, case_number <chr>,
#> #   sex <fct>, patient_origin <chr>, prescribed_zinc_supplement <fct>,
#> #   delivery_event <fct>, pregnancy_outcome_at_exit <fct>, date_of_exit <date>,
#> #   date_of_consultation_admission <date>, oedema <fct>, event_file_type <fct>,
#> #   residential_status_brief <fct>, pregnant <fct>, cholera_rdt_result <fct>,
#> #   dehydration_severity_during_stay <fct>, cholera_culture_result <fct>,
#> #   foetus_alive_at_admission <fct>, treatment_facility_site <int>,
#> #   prescribed_antibiotics <fct>, cholera_referred_to <fct>,
#> #   ors_consumed_litres <int>, treatment_location <chr>,
#> #   cholera_referred_from <fct>, cholera_treatment_facility_type <fct>,
#> #   date_of_onset <date>
#> > dict <- msf_dict(disease = "Cholera", long = TRUE, compact = FALSE, tibble = TRUE)
#> > print(dict)
#> # A tibble: 182 x 11
#>    data_element_uid data_element_na… data_element_sh… data_element_de…
#>    <chr>            <chr>            <chr>            <chr>           
#>  1 FF7d81Zy0yQ      egen_013_pregna… trimester        If pregnant, tr…
#>  2 FF7d81Zy0yQ      egen_013_pregna… trimester        If pregnant, tr…
#>  3 FF7d81Zy0yQ      egen_013_pregna… trimester        If pregnant, tr…
#>  4 ADfNqpCL5kf      egen_015_exit_s… exit_status      Final status of…
#>  5 ADfNqpCL5kf      egen_015_exit_s… exit_status      Final status of…
#>  6 ADfNqpCL5kf      egen_015_exit_s… exit_status      Final status of…
#>  7 ADfNqpCL5kf      egen_015_exit_s… exit_status      Final status of…
#>  8 ADfNqpCL5kf      egen_015_exit_s… exit_status      Final status of…
#>  9 ADfNqpCL5kf      egen_015_exit_s… exit_status      Final status of…
#> 10 wjCDTwXmtix      egen_064_treatm… treatment_facil… Name of facilit…
#> # … with 172 more rows, and 7 more variables: data_element_valuetype <chr>,
#> #   data_element_formname <chr>, used_optionset_uid <chr>, option_code <chr>,
#> #   option_name <chr>, option_uid <chr>, option_order_in_set <dbl>
#> > dict <- dict %>% select(option_code, option_name, data_element_shortname, 
#> +     option_order_in_set) %>% filter(!is.na(option_code))
#> > print(dict)
#> # A tibble: 165 x 4
#>    option_code option_name             data_element_shortname  option_order_in_…
#>    <chr>       <chr>                   <chr>                               <dbl>
#>  1 1           1st trimester           trimester                               1
#>  2 2           2nd trimester           trimester                               2
#>  3 3           3rd trimester           trimester                               3
#>  4 DH          Discharged home         exit_status                             1
#>  5 DD          Dead in facility        exit_status                             2
#>  6 DOA         Dead on arrival         exit_status                             3
#>  7 LA          Left against medical a… exit_status                             4
#>  8 AD          Transferred (to an MSF… exit_status                             5
#>  9 TR          Transferred (to Extern… exit_status                             6
#> 10 NO          None                    dehydration_level_at_a…                 1
#> # … with 155 more rows
#> > dat_clean <- clean_variable_spelling(dat, dict)
#> > print(dat_clean)
#> # A tibble: 20 x 45
#>    trimester exit_status treatment_facil… dehydration_lev… comments_on_lab…
#>    <fct>     <fct>       <lgl>            <fct>            <lgl>           
#>  1 NA        Left again… NA               Some             NA              
#>  2 NA        Dead on ar… NA               Severe           NA              
#>  3 NA        Discharged… NA               Some             NA              
#>  4 NA        Dead in fa… NA               Unknown          NA              
#>  5 NA        Dead on ar… NA               Severe           NA              
#>  6 NA        Transferre… NA               None             NA              
#>  7 NA        Dead on ar… NA               None             NA              
#>  8 NA        Transferre… NA               Unknown          NA              
#>  9 NA        Transferre… NA               Some             NA              
#> 10 NA        Transferre… NA               Unknown          NA              
#> 11 NA        Transferre… NA               Severe           NA              
#> 12 NA        Left again… NA               Unknown          NA              
#> 13 NA        Left again… NA               Some             NA              
#> 14 NA        Dead in fa… NA               None             NA              
#> 15 NA        Transferre… NA               Unknown          NA              
#> 16 NA        Discharged… NA               None             NA              
#> 17 NA        Transferre… NA               None             NA              
#> 18 NA        Transferre… NA               Some             NA              
#> 19 NA        Discharged… NA               Some             NA              
#> 20 NA        Dead in fa… NA               Unknown          NA              
#> # … with 40 more variables: fluids_treatment_plan <fct>, time_to_death <fct>,
#> #   malaria_rdt_at_admission <fct>, previously_vaccinated <fct>,
#> #   cholera_pcr_result <fct>, iv_fluids_received_litres <int>,
#> #   patient_origin_free_text <chr>, previous_vaccine_doses_received <fct>,
#> #   age_years <int>, arrival_date_in_area_if_3m <date>, age_days <lgl>,
#> #   msf_involvement <fct>, date_lab_sample_taken <date>, readmission <fct>,
#> #   age_months <int>, date_of_last_vaccination <date>, case_number <chr>,
#> #   sex <fct>, patient_origin <chr>, prescribed_zinc_supplement <fct>,
#> #   delivery_event <fct>, pregnancy_outcome_at_exit <fct>, date_of_exit <date>,
#> #   date_of_consultation_admission <date>, oedema <fct>, event_file_type <fct>,
#> #   residential_status_brief <fct>, pregnant <fct>, cholera_rdt_result <fct>,
#> #   dehydration_severity_during_stay <fct>, cholera_culture_result <fct>,
#> #   foetus_alive_at_admission <fct>, treatment_facility_site <chr>,
#> #   prescribed_antibiotics <fct>, cholera_referred_to <fct>,
#> #   ors_consumed_litres <int>, treatment_location <chr>,
#> #   cholera_referred_from <fct>, cholera_treatment_facility_type <fct>,
#> #   date_of_onset <date>