Unite estimates and confidence intervals

create a character column by combining estimate, lower and upper columns. This is similar to tidyr::unite().

unite_ci(x, col = NULL, ..., remove = TRUE, digits = 2,
  m100 = TRUE, percent = FALSE, ci = FALSE)

Arguments

x
a data frame with at least three columns defining an estimate, lower bounds, and upper bounds.

col
the quoted name of the replacement column to create

...
three columns to bind together in the order of Estimate, Lower, and Upper.

remove
if TRUE (default), the three columns in ... will be replaced by col

digits
the number of digits to retain for the confidence interval.

m100
TRUE if the result should be multiplied by 100

percent
TRUE if the result should have a percent symbol added.

ci
TRUE if the result should include "CI" within the braces (defaults to FALSE)

Examples

print(cfr <- case_fatality_rate((1:4)*10, 50))

#>   deaths population cfr    lower    upper
#> 1     10         50  20 11.24375 33.03711
#> 2     20         50  40 27.60839 53.81856
#> 3     30         50  60 46.18144 72.39161
#> 4     40         50  80 66.96289 88.75625

unite_ci(cfr, "CFR (CI)", cfr, lower, upper, m100 = FALSE, percent = TRUE)

#>   deaths population              CFR (CI)
#> 1     10         50 20.00% (11.24--33.04)
#> 2     20         50 40.00% (27.61--53.82)
#> 3     30         50 60.00% (46.18--72.39)
#> 4     40         50 80.00% (66.96--88.76)